+++ src/main/java/com/project/Ecommerce/Model/Product.java
@@ -1,11 +1,10 @@
 package com.project.Ecommerce.Model;
 
-
 import jakarta.persistence.*;
 import lombok.*;
 
 @Entity
-@Table(name="product")
+@Table(name = "product")
 @Getter
 @Setter
 @NoArgsConstructor
@@ -15,19 +14,18 @@ public class Product extends BaseEntity {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private  int id;
+    private Long id;
 
-    @Column(name = "name",nullable = false,length = 255)
-    private  String name;
+    @Column(name = "name", nullable = false, length = 255)
+    private String name;
 
     private Float price;
 
-    @Column(name = "thumbail",length = 300)
+    @Column(name = "thumbail", length = 300)
     private String thumbnail;
 
     @Column(name = "description", columnDefinition = "LONGTEXT")
     private String description;
-
 
     @ManyToOne
     @JoinColumn(name = "categories_id")
